Added comments to post_routes and user_routes files

+++ blogapp/user_routes.py
@@ -13,20 +13,51 @@ def set_session(user_details):
     session["profile_picture"] = user_details.profile_picture
     return None
 
+def remove_session():
+    session.pop('user_id', None)
+    session.pop('name', None)
+    session.pop('username', None)
+    session.pop('profile_picture', None)
+    return None
+
+def thumbnail_profile_picture(profile_pic):
+    output_size = (200, 200)
+    profile = Image.open(profile_pic)
+    profile.thumbnail(output_size)
+    return profile
+
+def save_picture(profile_pic, filename):
+    file_ext = os.path.splitext(filename)[1]
+    new_filename = str(session["user_id"])+ file_ext
+    picture = thumbnail_profile_picture(profile_pic)
+    picture.save(os.path.join(app.config["UPLOAD_PATH"], new_filename))
+    return new_filename
+
+def send_reset_email(user):
+    token = user.get_reset_token()
+    msg = Message('Password Reset Request', 
+                sender='no.reply.mendonca@gmail.com',
+                recipients = [user.email])
+    msg.body =f'''To reset your password, visit the following link:
+{url_for('reset_token', token = token, _external=True)}
+If you don't receive this email in one minute, please check spam folder.
+If you didn't request this, please ignore this email.  
+'''
+    mail.send(msg)
+
+# App routes
+
+# Homepage route
 @app.route("/")
 def homepage():
     posts = db.session.execute(db.select(Post).order_by(Post.postedOn.desc())).scalars()
     return render_template("homepage.html", title="Blog Homepage", posts = posts)
 
-@app.route("/<int:id>")
-def user_posts(id):
-    posts = db.session.execute(db.select(Post).where(
-            Post.user_id == id).order_by(Post.postedOn.desc())).scalars()
-    return render_template("posts.html", title="Blog Homepage", posts = posts)
-
 @app.route("/login", methods=['GET', 'POST'])
 def login():
     if request.method == "POST":
+
+        # Queries the database to find an user with a specific email
         userExists =  db.session.execute(db.select(User).where(
             User.email == request.form["email"])).scalar()
         if userExists:
@@ -34,14 +65,16 @@ def login():
                 User.email == request.form["email"])).scalar()
             formPassword = request.form["password"]
 
+            # Checks if user introduced a valid password
             if bcrypt.check_password_hash(userPassword, formPassword):
                 user = db.session.execute(db.select(User).where(
                     User.email == request.form["email"])).scalar()
+                
+                # Adds user details to session
                 set_session(user)
                 return  redirect(url_for('homepage'))
         flash("Email or password invalid, please try again.")
     return render_template("login.html", title="login")
-
 
 @app.route("/signup", methods=['GET', 'POST'])
 def signup():
@@ -49,16 +82,21 @@ def signup():
         if request.form["password"] != request.form["confPassword"]:
             flash("Passwords don't match")
         else:
+            # Checks if username already taken
             usernameExist = db.session.execute(db.select(User).where(
                 User.username == request.form["username"].lower())).scalar()
             if usernameExist:
                 flash("Username not available")
             else:
+
+                # Checks if user has already an account with that email
                 emailExist = db.session.execute(db.select(User).where(
                     User.email == request.form["email"])).scalar()
                 if emailExist:
                     flash("Email already exists")
                 else:
+
+                    # Creates a new user account
                     user = User(
                         name=request.form["name"],
                         username=request.form["username"].lower(),
@@ -72,53 +110,68 @@ def signup():
                     return redirect(url_for('homepage'))
     return render_template("sign_up.html", title="signup")
 
+# route to posts by specific user
+@app.route("/user/<int:id>")
+def user_posts(id):
+    posts = db.session.execute(db.select(Post).where(
+            Post.user_id == id).order_by(Post.postedOn.desc())).scalars()
+    return render_template("posts.html", title="Blog Homepage", posts = posts)
 
+# Removes user details from session
 @app.route("/logout")
 def logout():
-    session.pop('user_id', None)
-    session.pop('name', None)
-    session.pop('username', None)
-    session.pop('profile_picture', None)
+    remove_session()
     return redirect(url_for('homepage'))
 
-
+# Renders account page
 @app.route("/account")
 def account():
     return render_template("account.html")
 
+# Allows user to upload a profile picture 
 @app.route("/upload_file", methods=['GET', 'POST'])
 def upload_file():
     if request.method == "POST":
         profile_pic = request.files['profile_picture']
+        # Returns a secure filename
         filename = secure_filename(profile_pic.filename)
+        if filename == '': 
+            flash('Please upload a picture with a valid name')
+            return redirect(url_for('account'))
+
+        # Removes previous profile picture from folder
         if session["profile_picture"] != "default.jpeg":
             os.remove(f"blogapp/static/uploads/{session['profile_picture']}")
-        file_ext = os.path.splitext(filename)[1]
-        new_filename = str(session["user_id"])+ file_ext
-        output_size = (200, 200)
-        profile = Image.open(profile_pic)
-        profile.thumbnail(output_size)
-        profile.save(os.path.join(app.config["UPLOAD_PATH"], new_filename))
-        db.session.execute(db.update(User).values(profile_picture = new_filename).where(
+        
+        # saves picture to folder and returns filename
+        picture_name = save_picture(profile_pic, filename)
+
+        # saves filename to database
+        db.session.execute(db.update(User).values(profile_picture = picture_name ).where(
             User.id == session["user_id"]))
         db.session.commit()
-        session["profile_picture"] = new_filename
+
+        # updates session with new profile picture details
+        session["profile_picture"] = picture_name
         return redirect(url_for('account'))
     return render_template("account.html")
     
-
+# Updates user details in the database
 @app.route("/update", methods=['GET','POST'])
 def update_details():
     if request.method == "POST":
-        if request.form["name"] == '' and request.form["username"] == '':
+        # If user doesn't update name or username, it flashes an error message
+        if request.form["name"] == "" and request.form["username"] == "":
             flash("Please provide name/username")
             return redirect(url_for('account'))
-        elif request.form["username"] == '':
+        
+        # updates the name
+        elif request.form["username"] == "":
             db.session.execute(db.update(User).values(name = request.form["name"]).where(
             User.id == session["user_id"]))
-            db.session.commit()
             session["name"] = request.form["name"]
-        else:
+        elif request.form["name"] == "":
+            # Updates username if not taken
             usernameExist = db.session.execute(db.select(User).where(
                 User.username == request.form["username"].lower())).scalar()
             if usernameExist:
@@ -126,42 +179,49 @@ def update_details():
             else:
                 db.session.execute(db.update(User).values(username = request.form["username"].lower()).where(
                     User.id == session["user_id"]))
-                db.session.commit()
                 session["username"] = request.form["username"]
-        return redirect(url_for('account'))
+        else:
+            # updates name and username in database
+            db.session.execute(db.update(User).values(name = request.form["name"], 
+                username = request.form["username"].lower()).where(
+                User.id == session["user_id"]))
+            session["name"] = request.form["name"] 
+            session["username"] = request.form["username"]
+
+        # Adds changes to databse and redirects to account page
+        db.session.commit()
+        return redirect(url_for('account'))   
     return render_template("account.html")
 
-def send_reset_email(user):
-    token = user.get_reset_token()
-    msg = Message('Password Reset Request', 
-                sender='no.reply.mendonca@gmail.com',
-                recipients = [user.email])
-    msg.body =f'''To reset your password, visit the following link:
-{url_for('reset_token', token = token, _external=True)}
-If you didn't request this, please ignore this email.  
-'''
-    mail.send(msg)
-
+# route to reset password email form
 @app.route("/reset_password", methods=['GET','POST'])
 def reset_request():
     if request.method == 'POST':
         user =  db.session.execute(db.select(User).where(User.email == request.form["email"])).scalar()
-        send_reset_email(user)
-        flash("An email with instructions have been sent.")      
-        return redirect(url_for('login'))
+        if user:
+            send_reset_email(user)
+            flash("An email with instructions have been sent.")      
+            return redirect(url_for('login'))
+        # flashes an error if there isn't an user with that account
+        flash("Invalid email")
     return render_template('reset_request.html', title = 'Reset Password')
 
+# route to reset password from email
 @app.route("/reset_password/<token>", methods =['GET', 'POST'])
 def reset_token(token):
+    # verify_reset_token checks if the token is valid and returns an user if it is
     user = User.verify_reset_token(token)
     if user is None:
         flash('That is an invalid or expired token')
         return redirect(url_for('reset_request'))
+    
     if request.method == 'POST':
+        # it flashes an error if passwords don't match
         if request.form["password"] != request.form["confPassword"]:
             flash("Passwords don't match")
             return redirect(url_for('reset_request'))
         else:
+            # Encrypts password and saves it to database
             password=bcrypt.generate_password_hash(
                     request.form["password"]).decode('utf-8')
             user.password = password
@@ -170,54 +230,4 @@ def reset_token(token):
             return redirect(url_for('login'))
     return render_template("reset_token.html", title = 'Reset Password')
 
-
-@app.route("/post", methods=['GET', 'POST'])
-def post():
-    if request.method == "POST":
-        post = Post(
-            title = request.form["title"],
-            text = request.form["post"],
-            user_id = session["user_id"],
-        )
-        db.session.add(post)
-        db.session.commit()
-        return redirect(url_for('homepage'))
-    return render_template("homepage.html", title="Homepage")
-
-@app.route("/newpost")
-def newpost():
-    return render_template("new_post.html", title="New post")
-
-@app.route("/post/<int:id>")
-def singlepost(id):
-    post = db.session.execute(db.select(Post).where(
-            Post.id == id)).scalar()
-    return render_template("post.html", post = post)
-
-@app.route("/delete/post/<int:id>", methods=['GET','POST'])
-def delete_post(id): 
-    db.session.execute(db.delete(Post).where(Post.id == id))
-    db.session.commit()
-    flash("Post deleted")
-    return redirect(url_for('homepage'))
-
-@app.route("/post/updateform/<int:id>", methods=['GET', 'POST'])
-def updateform(id):
-    post = db.session.execute(db.select(Post).where(Post.id == id)).scalar()
-    return render_template("update_post.html", post= post)
-
-@app.route("/update/post/<int:id>", methods = ['GET','POST'])
-def updatepost(id):
-    if request.method == 'POST':
-        if request.form['title'] =='' and request.form['post']== '':
-            return redirect(url_for('singlepost', id = id))
-        elif request.form['title'] == '':
-            db.session.execute(db.update(Post).values(text = request.form["post"]).where(Post.id == id))
-        elif request.form['post'] == '':
-            db.session.execute(db.update(Post).values(title = request.form["title"]).where(Post.id == id))
-        else:
-            db.session.execute(db.update(Post).values(title = request.form["title"],text = request.form["post"] ).where(Post.id == id))
-        db.session.commit()  
-        return redirect(url_for('singlepost', id = id))
-    return redirect(url_for('updateform', id = id))
